thoughtfulLM/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
httpx==0.25.0
python-dotenv==1.0.0
openai>=1.0.0
# thought_kit will be installed in development mode (editable) 